Update patch set 2
Patch Set 2:

I think this change is not good as it is somewhat circular. It assumes that if a display was found that there is a display and if none was found that there is none. I would move this code into an explicit test (comparing with lsb-release as below).

The hardware dependency should be keyed off /etc/lsb-release saying CHROMEBIT or CHROMEBOX (no display), otherwise yes.

Patch-set: 2