Update patch set 3
Patch Set 3:

The code is reasonable, it checks if the kernel found a display. But I think there may be inconsistencies between what the lab scheduling thinks a device with a display is, what the kernel thinks a device with a display is. (Or maybe it is simply not wired up yet?) Now for some features discrepancy makes sense, hardware could be disabled in some branches for instance. Surprised that we need it here.

Patch-set: 3
Reviewer: Gerrit User 1000221 <1000221@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>