Update patch set 12
Patch Set 12:

Updated commit message with the correct version number.

For testing, deps_test.go (which was revised by crrev/c/2117371) was one of the test that is related to this change. It is one of the tests that is run by fast_build.sh -T as noted in TEST=.
Do you mean it needs a different kind of test or should be more specific?

Patch-set: 12
Reviewer: Gerrit User 1166720 <1166720@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>