Update patch set 1
Patch Set 1:

I think it's better to land this after we have R arcvm for both x86 and arm,

if we just launch this for now, that means we don't have coverage on arm arcvm.

If we don't really care about eve CQ then we can just change back all x86 P dual
build boad to a single arc++ board if we don't care arcvm-p on it.

Patch-set: 1