Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> i don't understand the copies here
> 
> /config is already branched and can be cros_workon ref'd from anywhere

Right, this patch was a tentative change when some messages that we wanted use resided inside /infra, and some message types in it were referring another message type in /config (the script in that repository was intended to compile combine those files across repositories upon compilation). We can remove this change now after switching to use the branched /config.

Patch-set: 9