Update patch set 1
Patch Set 1:

Looking into this it is not clear as to why this is failing. I think this will take a while to fix, so feel free to add auron to the unstable boards list.  guado does not appear to be running on the CQ though.

You may also need to change the tast-use-flags ebuild as it lists auron_yuma and auron_paine which I don't believe are declared by the variant overlays.

Patch-set: 1
CC: Gerrit User 1370239 <1370239@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>