tast: Allow specifying extra allowed GCS buckets.
In partner environment it is possible that data files are in
different buckets than what Google uses, but access to them are
rejected due to the hard-coded allowlist in the ephemeral devserver.

This change adds -extraallowedbuckets flag to the tast command to
allow specifying extra buckets.

BUG=chromium:1069786
TEST=fast_build.sh -T

Change-Id: I4682f90e400d428c3e67d77dd89dffb73b230013
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2164288
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>