Add "crash_syslog_(un)stable" deps.
Per the linked bugs, some crash tests are broken on 5.4 kernels. The
root cause is not well-understood yet, but we believe that it's a
legitimate issue with the kernel (rather than a test bug).

We want to un-mark these tests as critical on these boards to allow lab
development to proceed.

BUG=b:155009503,chromium:1072054
TEST=fast_build.sh

Change-Id: I900ba2b926dd5821a8192068dce12bead83b9340