dut: Add comparable error for companionDeviceHostname.
When running WiFi tests in a local setup, it is possible that pcap
device is not available. It will be convenient if we can fallback to
use test router instead when:
- default pcap is used
- hostname of default pcap is invalid

Make the error of companionDeviceHostname comparable so that we can
detect the case from its caller.

This behavior will be used in crrev.com/c/2172561.

BUG=chromium:1034875
TEST=fast_build.sh -T

Change-Id: If57240fbf99fc36f2fe7d5b171b9fe68667044a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2172110
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>