tast: Associate PreCtx with the first test.
Historically the context.Context and testing.State given to
preconditions are those for the first test. For example,
testing.State.ContextOutDir returns the output directory for the
first test. This is known to be a suboptimal design, but it is
the current semantics.

Currently the context returned by testing.State.PreCtx is not
associated with any test, meaning that it cannot be used to call
support libraries gated with current test checks, e.g. gRPC method
calls (w/ testing.ContextServiceDeps) or chrome.New
(w/ testing.ContextSoftwareDeps). This change associates it with
the first test to be consistent with the existing semantics.

BUG=chromium:1079167
TEST=fast_build.sh -T

Change-Id: I9a9cb360d49d7277a649c795605e821736ea23fc