tast: Unified logging via context.
Formerly logging via context was possible only in tests where
TestContext is associated with context.Context. This prevented framework
code from using the same logging mechanism.

This change introduces chromiumos/tast/internal/logging package for the
unified logging via context.Context. Now testing.ContextLog is a simple
wrapper of logging.ContextLog, and all test contexts are associated with
a log sink by logging.NewContext. Consequently, logger is no longer a
part of TestContext, so the code around TestContext are simplified.

This change focuses on introducing the new logging package and
redirecting tests to use it via public API in the testing package.
Refactoring of larger framework code to use the new logging package
will be done in a separate change.

BUG=chromium:1081562
TEST=fast_build.sh -T

Change-Id: Iebe817587444967f660b11d10f7ec088c3eecdcc