tast: Introduce the dep package.
This is a part of the patch series to do software/hardware
dependency checks in a consistent manner.

This patch moves chromiumos/tast/internal/testing/hwdep to
chromiumos/tast/internal/dep with some renaming.

The new package will be the central package handling all test
dependencies, including software dependencies. For now, only
hardware dependencies are handled to minimize the diff size.

This is a refactoring change; no functional change is expected.

BUG=chromium:1077090
TEST=fast_build.sh -T

Change-Id: Ib3be1b3aff2a9873fb6834b9250f39f5c55fbc59