tast: Introduce dep.Features struct.
This is a part of the patch series to do software/hardware
dependency checks in a consistent manner.

This patch introduces dep.Features struct which holds information of
both software/hardware features of the DUT. Software features are
represented by dep.SoftwareFeatures struct (formerly
runner.SoftwareFeatures), and hardware features are represented by
dep.HardwareFeatures struct (the former hwdep.DeviceConfig
struct.

TestInstance.ShouldRun is modified to accept Features as the only
argument.

This is a refactoring change; no functional change is expected.

BUG=chromium:1077090
TEST=fast_build.sh -T

Change-Id: I3b933e2eba7b9fe4a7de5db6bfb668ed6b444a3a