tast: Deprecate TestEnd.MissingSoftwareDeps field.
This is a part of the patch series to do software/hardware
dependency checks in a consistent manner.

We are going to refactor test bundles so that missing software
features information is not returned from the testing package.
This change removes one use case of returned missing software
features information. To notify skips, we can still use the existing
field for unsatisfied hardware dependencies.

BUG=chromium:1077090
TEST=fast_build.sh -T

Change-Id: I1d12107f5024d3d9171d3482a80f3028e8698def
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2209947
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>