tast: Check unknown software dependencies in ShouldRun.
This is a part of the patch series to do software/hardware
dependency checks in a consistent manner.

Formerly software dependency checks were scattered to two places:
TestInstance.ShouldRun which checks if dependencies are satisfied,
and reportSkippedTest which checks if failed software dependencies
were due to unknown software features.

To perform the latter check, SkipReason struct returned by ShouldRun
used to contain a list of missing software features. But for consistent
software/hardware dependency checks, we do not want to have the field.

This patch moves the unknown software feature checks from
reportSkippedTest to ShouldRun. Then ShouldRun no longer need to
report missing software features in a special way.

This is a refactoring change; no functional change is expected.

BUG=chromium:1077090
TEST=fast_build.sh -T

Change-Id: I4f6a62853b425d6b81c03263e133505e4ca2e3cb