tast: Add chrome_media as a software dependency to handle proprietary codec
Chrome_internal is a USE flag that can be built with only internally
by Google. As a partner, when we build Chromium OS images to test
proprietary codecs such as H.264 we set USE=”chrome_media”.
These are focused development images and not distributed.
However, some tests like arc.LibvdaDecde.h264 depend on chrome_internal
as a software dependency. This test checks video decoding using libvda's
Mojo connection to GAVDA is working. We enable the intel-media-driver
on X86 platforms and hence need to test H.264 codecs.

The goal is to have a more inclusive software dependency for proprietary
codec tests by using chrome_media instead of chrome_internal.
We suggest this new dependency to be named chrome_media. So that this
could be tested on Chromium OS images built without the chrome_internal
USE flag

BUG=chromium:1084463
TEST=tast run <DUT IP> arc.LivdaDecode.h264

Cq-Depend:chromium:2276739
Change-Id: Ibf2fe06472225dd9624f66a2531dae5923dfd30b
Signed-off-by: Madhulika <madhulika.dolke@intel.com>