tast: Move dependency check logic to the dep package.
This is a part of the patch series to do software/hardware
dependency checks in a consistent manner.

This patch moves the dependency check logic from ShouldRun to
dep.Check. Two new types, dep.Deps and dep.SoftwareDeps, are
introduced to help ShouldRun to delegate checks to dep.Check.

This is a refactoring change; no functional change is expected.

BUG=chromium:1077090
TEST=fast_build.sh -T

Change-Id: I15fbfdc3a0a8b3209ee239f95f482e37d06f90d6