planner: Extract TestInstance.Run to a function in a separate file.
This is a part of the patch series to extract the test execution
logic out of the testing package to the planner package.

This is a pure code move to prepare to minimize the diff in the later
changes. No functional change is expected.

BUG=chromium:1077090
TEST=fast_build.sh -T

Change-Id: I5382df649f50761f6fe17e5d516dd7e5291653a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2217705
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>