planner: Prepare for the big move.
This is a part of the patch series to extract the test execution
logic out of the testing package to the planner package.

This change exports some symbols to prepare for the following
move of the Run function to another package.

output_test.go is renamed to output.go to allow accessing it from
the new package. This is suboptimal, but I'm planning to remove
Output very soon.

This is a refactoring change; no functional change is expected.

BUG=chromium:1077090
TEST=fast_build.sh -T

Change-Id: I94f560e50ad2486bbc1db447a134b1e72128e095