planner: Add rootState.hasError method.
This is a part of the patch series to extract the test execution
logic out of the testing package to the planner package.

Just like State, rootState also keeps track of reporter errors
and allow checking if any error was reported. It is used in
TestInstance.Run.

This is a refactoring change; no functional change is expected.

BUG=chromium:1077090
TEST=fast_build.sh -T

Change-Id: Ic1ef08be0ea86af5c375d29ef26aabc717661e58