planner: Explicitly construct rootState.
This is a part of the patch series to extract the test execution
logic out of the testing package to the planner package.

Instead of constructing State directly, ask callers to construct
rootState first and derive State from it.

This is a refactoring change; no functional change is expected.

BUG=chromium:1077090
TEST=fast_build.sh -T

Change-Id: If07779bca3e31fbf51d24019e8c08ccecc995320
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2217702
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>