planner: Make State.inPre immutable.
This is a part of the patch series to extract the test execution
logic out of the testing package to the planner package.

Instead of changing the fields of State, create a new State for
preconditions.

This is a refactoring change; no functional change is expected.

BUG=chromium:1077090
TEST=fast_build.sh -T

Change-Id: If6a9a405533037d6886b881ccb4366bd96453c94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2217701
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>