planner: Introduce a fake precondition for non-execution tests.
This is a part of the patch series to extract the test execution
logic out of the testing package to the planner package.

Introduce fakePre, an empty precondition, to be used for unit tests
that want to declare a precondition but do not care its implementation.

BUG=chromium:1077090
TEST=fast_build.sh -T

Change-Id: If46db373ccd40d6692ac3dd329851765100e60a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2217980
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>