bundle: Abort execution if a test ignores its timeout.
Today, if a test main function does not return soon after the context
cancellation by the test timeout, Tast just lets the goroutine running
and continue to the next test.

We found that this is a problematic behavior because an abandoned
goroutine may interact the next test badly, e.g. calling chrome.Close
concurrently with chrome.New.

Rather than continuing, we should abort the bundle execution when a test
does not return. Fortunately the tast command has logic to restart test
execution when runners/bundles exit before all tests are run, so we can
continue from the next test.

BUG=chromium:1087758
TEST=fast_build.sh -T
TEST=tast run DUT example.Block example.Pass

Change-Id: I51b62d0a5a606dff7ca079cf467ea20bd1d9bb62