tast: Make -continueafterfailure=true default.
The flag is for retrying a runner execution when it exits prematurely,
for example due to test crashes. CI systems always set this flag to
get test results as much as possible and its usefulness was proved.
Let us make the flag on by default.

BUG=chromium:1087758
TEST=fast_build.sh -T

Change-Id: I980c25fd734c5ff83572164175560c043bb96d81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2217982
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>