tast: hwdep: Introduce hwdep.[No]ForceDischarge
This CL introduces hwdep.[No]ForceDischarge to enable checking if
the device can set its battery to discharge state for testing when
the AC is connected. These hwdeps will be used to prevent setting
up the device to discharge state which will be failed on the devices
which do not support ectool chargecontrol.

BUG=b:157512297, b:153027513
TEST=tested with chromium:2237579

Change-Id: I6537bdb4819279c5cb2db8b5c33933c5e6699e57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2228923
Tested-by: Hikaru Nishida <hikalium@chromium.org>
Commit-Queue: Hikaru Nishida <hikalium@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>