tast: hwdep: Introduce hwdep.[No]ForceDischarge
This CL introduces hwdep.[No]ForceDischarge to enable checking the
device can set its battery to discharge state for testing even if
the AC is connected. These hwdeps will be used to block setting up
the device to discharge state which will be failed on the devices
which do not support ectool chargecontrol.

BUG=b:157512297, b:153027513
TEST=tested with chromium:2237579

Change-Id: I6537bdb4819279c5cb2db8b5c33933c5e6699e57