Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > https://chromium.googlesource.com/chromiumos/infra/proto/+/refs/heads/master/src/device/config.proto#125
> > 
> > could you talk to infra team and extend the protobuf, first,
> > then, import it to Tast, and add some logic to detect the SoC?
> 
> To me it sounds as if you're speaking Greek.  Huh?  I have no idea what any of that means.

The device/config.pb.go is generated from a .proto file.
The original proto file is maintained in the repository I linked.
So, in order to add a new enum value, please add it in the repository first, which is maintained by the chromeos infra team.
The files in tast repository delivered from it can be updated, then.

Background: Tast as a framework needs to provide test metadata to infra for scheduling, so the schema needs to be shared. What you're changing is the schema.
In the future, the actual device data will be provided from infra. Because it'd take longer time, for short term, we agreed to "detect" it on runtime in Tast, and that's why we have such a code in features.go which you're editing.

Patch-set: 2
Reviewer: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>