tast: Replace Output channel with OutputStream interface.
This change replace the way to report test outputs by the new
OutputStream interface. The interface is implemented by
testEventWriter.

Since we don't use channels to deliver test outputs, the copier
goroutine is no longer needed and thus removed.

Although most of the current behavior is maintained, there is a
very subtle change to the timing where the output stream is closed.
Formerly the output channel was closed when a test function returned.
This was actually problematic because, if a test writes logs after
the copier goroutine exits due to a test timeout, the output sent to
the channel was not read by anyone and thus the test function blocked
forever. In contrast, OutputStream does not have a Close method,
instead it is closed when planner.Run returns. So after a test timeout,
logs emitted by the test are silently dropped without blocking.

BUG=chromium:1091601
TEST=fast_build.sh -T

Change-Id: I10e0898d7b5ba3152bcdbd8089b1416ff5700568
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2230751
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>