planner: Increase the timeout to wait for test functions to return.
Tests should return immediately when a test timeout is reached and
the context is canceled. However, in reality, tests sometimes take
a while to return, e.g. if it tries to copy large files in deferred
function calls. And when it happens tests may not report useful error
messages.

Currently only 3 seconds is allowed between a test timeout and test
bundle abort. This patch increases it to 30 seconds.

BUG=chromium:1087058
TEST=fast_build.sh -T

Change-Id: I42b092e2a6a755752556a370661be6a70d5c799b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2233430
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>