tast: Introduce PreState type.
PreState is a State type for preconditions. It is initially an alias
of State, but will diverge from State once users are updated.

baseState type is introduced to attach methods common to State and
PreState, and it is embedded to State and PreState.
Test/precondition-specific methods are attached to State/PreState
directly, e.g. State.PreValue and PreState.PreCtx.

BUG=chromium:1090644
TEST=fast_build.sh -T

Change-Id: I7cc826d184b00e827eef57889579b4d69ce64712