tast: Introduce PreState type.
PreState is a State type for preconditions. It is initially an alias
of State, but will diverge from State once users are updated.

baseState type is introduced to attach methods common to State and
PreState, and it is embedded to State and PreState.
Test/precondition-specific methods are attached to State/PreState
directly, e.g. State.PreValue and PreState.PreCtx.

If a function wants to take both State and PreState, it should
define an interface. As for testing.NewContext, StateForContext
interface is added (but it is not exposed to tests).

BUG=chromium:1090644
TEST=fast_build.sh -T

Change-Id: I7cc826d184b00e827eef57889579b4d69ce64712
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2237553
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>