tast: hwdep: Add condition for WiFi mac randomization support.
Some tests on WiFi MAC randomization are about to be added. Add a
condition for the dependency. As we don't yet have the hw condition on
WiFi chips, create the condition with skip list for now. Reasons are
stated in comments so we can migrate it to hwdep on WiFi chip later.

BUG=b:158061732
TEST=fast_build.sh -T

Change-Id: I516c5d63132a6060e7581b0b7c9b2ba89002b964
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2240937
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>