tast: hwdep: Add condition for WiFi mac randomization support.
Some tests on WiFi MAC randomization are about to be added. Add a
condition for the dependency. As we don't yet have the hw condition on
WiFi chips, create the condition with skip list for now. Reasons are
stated in comments so we can migrate it to hwdep on WiFi chip later.

BUG=b:158061732
TEST=fast_build.sh -T
TEST=Run the test in crrev.com/c/2228285 and make sure the specific
platforms are skipped.

Change-Id: I516c5d63132a6060e7581b0b7c9b2ba89002b964