dut: allow for exporting Hostname
In order to run syzkaller against a DUT from within a tast wrapper, the
wrapper needs to provide syzkaller with information related to the SSH
connection(hostname, port, SSH keyfile).

Currently we have (*DUT).KeyFile() to retrieve the path to the SSH key;
so add (*DUT).HostName() to retrieve "<hostname>:<port>" information.

BUG=b:156755711
TEST=fast_build.sh -T

Change-Id: I1adc008949361b58caa723efbecaa342c6a35a63