Reland: tast: Replace Output channel with OutputStream interface.
This change replace the way to report test outputs by the new
OutputStream interface. The interface is implemented by
testEventWriter.

Since we don't use channels to deliver test outputs, the copier
goroutine is no longer needed and thus removed.

Although most of the current behavior is maintained, there is a
very subtle change to the timing where the output stream is closed.
Formerly the output channel was closed when a test function returned.
This was actually problematic because, if a test writes logs after
the copier goroutine exits due to a test timeout, the output sent to
the channel was not read by anyone and thus the test function blocked
forever. In contrast, OutputStream does not have a Close method,
instead it is closed when planner.Run returns. So after a test timeout,
logs emitted by the test are silently dropped without blocking.

The original CL (crrev.com/c/2230751) had a data race in unit tests.
This reland fixes the issue.

BUG=chromium:1091601
TEST=fast_build.sh -T
TEST=fast_build.sh -C
TEST=go.sh test -count 1000 chromiumos/tast/internal/testing

Change-Id: I874c20c76066598357f8cd7aa62d90096dfb8dc0