tast: Remove -readfile option from "tast list".
There should be no user of the option actually. It was added to be
used by chromite for hardware dependency support (crbug.com/950347),
but the approach changed a lot since then.

Rather the feature uses TestInstance in a weird way and can be an
obstacle for the upcoming introduction of TestInfo.

BUG=chromium:950347
BUG=chromium:1091601
TEST=fast_build.sh -T
TEST=codesearch tast "-readfile"

Change-Id: I8473279b58fb17c7b5029faac2c18c302fa80952