tast: Convert TestInstance.DataDir to a function RelativeDataDir.
After recent refactorings of TestInstance, DataDir still being a method
of TestInstance is a bit odd. Also we want to compute DataDir from an
upcoming TestInfo. Thus this change extracts the method as a normal
function.

This is a pure refactoring; no functional change is expected.

BUG=chromium:1091601
TEST=fast_build.sh -T

Change-Id: Id452c72497f7f4961166fb4e120e8474ae9da356
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2253824
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>