tast: Introduce testing.TestInfo.
Formerly TestInstance had two purposes: (1) to represent a test
instance registered to the global registry in bundles, (2) to
send test information over wire among the tast command, runners and
bundles.

But this was known to be a tricky design because serializing and
deserializing TestInstance does not yield the original TestInstance
due to existence of runtime-only fields (e.g. Func, Pre, Val).

This change finally solves the issue by introducing a small struct
TestInfo dedicated for the second purpose of the original TestInstance.
TestInfo is a simple struct consisting of JSON-serializable exported
fields only. TestInstance is no longer JSON-serializable.

Users of TestInstance are updated to use TestInfo instead whenever
appropriate. In particular, control messages now deliver TestInfo
instead of TestInstance.

This is a pure refactoring; no functional change is expected.

BUG=chromium:1091601
BUG=chromium:984387
TEST=fast_build.sh -T

Change-Id: I3a0493010aa76643add30f7d65f6384199bd1bf7