tast: Remove SSH command announcement checks.
Unit tests in chromiumos/tast/cmd/tast/internal/run used to check that
an SSH command sent from the tast command exactly matches with the one
received by a fake SSH server. This was done by having a hook in the
ssh package to "announce" a command being run.

This was not a very useful check because it essentially just tests the
underlying SSH library but nothing else. Furthermore, the announcement
mechanism was not goroutine-safe, which made unit tests flaky when
multiple files were copied in parallel.

This patch removes the announcement checks entirely. The fake SSH
server is updated to actually run a requested command by default.

BUG=chromium:1086335
TEST=fast_build.sh -T
TEST=go.sh test -count 100 -run TestLocalCopyOutput \
     chromiumos/tast/cmd/tast/internal/run

Change-Id: I8c28dcf22f3115d536297b7a40d26aef8f1f4097
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2262652
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>