testing: Remove SortTests.
SortTests used to give the exact order of test execution, but it is
no longer true.

This patch removes the function. The only caller in the bundle package
instead uses a simple sort by test names. This changes the order of
test enumeration in "tast list", but there was no such guarantee from
the beginning, and no one depends on the order as far as I know.

BUG=chromium:1091601
TEST=fast_build.sh -T
TEST=tast list DUT

Change-Id: I3f739d1995ecd0e6adc15ac132fe4e49bf03272e