bundle: Rename preTestFunc to testHook.
preTestFunc is called before tests, but it can also return a function
to be called after tests. To avoid confusion, this patch renames it to
testHook.

Fields in bundle.LocalDelegate is renamed later because changing it
requires updating the caller as well.

No functional change expected.

BUG=chromium:965703
TEST=fast_build.sh -T

Change-Id: I4f4e52df91cc7ab6d8adba0af2cc18c479adf163
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2287012
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>