tast: Introduce testing.TestHookState.
We want to provide a few more methods to test hooks, while we want to
hide some State-specific methods (Param, PreValue, Run). Thus it is
high time to introduce a new State type for test hooks.

This change has to be atomic with a bundlemain change.

No functional change is expected.

BUG=chromium:965703
BUG=chromium:1103118
TEST=fast_build.sh -T

Cq-Depend: chromium:2286973
Change-Id: Ia97839d0d775b9a2147542f0147e3d5c601d7c57