tast: Provide purgeable file paths to test hooks.
testing.TestHookState.Purgeable returns a list of file paths which are
not used for now and thus can be deleted if the disk space is low.
Initially it contains a list of external data files that are not used
in the current test plan.

This allows test hooks to delete large files to make room for the next
test.

This change adds a framework API.

BUG=chromium:965703
TEST=fast_build.sh -T

Change-Id: I127ea546543acdbbf2974f9146b41ceb694cf9ae