tast: Allow configuring strategy to download external data files
This change adds a new flag -downloaddata={batch,lazy} to the Tast CLI
which specifies the strategy to download external data files.

-downloaddata=batch (default)

  Downloads all relevant external data files in batch before running
  tests. This is the current behavior.

-downloaddata=lazy

  Downloads relevant external data files as needed between tests.
  This is a new behavior to minimize local disk usage.

This configuration value is sent to test bundles in bundle.RunTestsArgs.
In the end, the value is consumed by the planner package to decide when
to download external data files.

Protocol compatibility consideration:
- New Tast CLI vs. older test bundles: -downloaddata is ignored.
- Old Tast CLI vs. newer test bundles: -downloaddata=batch is assumed.

BUG=chromium:965703
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=~/go/bin/tast run -downloaddata=lazy DUT meta.LocalFail meta.LocalFiles

Change-Id: Ifb9fac2b9a4662781a9603f072f112a2326c6e6c