hwdep: Add WifiNotMarvell condition.
We have some known test failure for Marvell WiFi like b/158150763 which
might not be resolved in short term (e.g. need firmware fix). In this
case, we would like to temporary skip running the tests on them so that
the tests on other platform can be promoted to stable suites to be
better tracked.

As we don't yet have a field in device.Config for WiFi chip information,
add a hwdep.Condition by listing the platforms for this specific usage
for now.

BUG=b:158150763
TEST=fast_build.sh -t chromiumos/tast/testing/hwdep

Change-Id: Ib391eda5af15fe21d62f17d3216018ebc229f2d9