Launch crashpad to 100% on most boards
There are currently three classes of boards:
1. Most boards, which are running an experiment where Chrome is run 10%
   of the time with Crashpad and 90% of the time with Breakpad.
2. kukui, cheza, trogdor, and asurada, which run Crashpad 100% of time
3. scarlet, bob, kevin, rainier, oak, elm, and hana, where Crashpad
   doesn't work so we run Breakpad 100% of the time

We have finished the experiment and are ready to launch Crashpad to 100%
on most boards.

These changes switch the boards in category 1 (the Crashpad/Breakpad
experiment group) to running Crashpad 100% of the time.

Before, since there were 3 categories, we represented the categories
using two USE flags: force_crashpad and force_breakpad. If
force_crashpad was set, the board was in category 2. If force_breakpad
was set, the board was in category 3. If neither was set, the board was
in category 1.

Now, there are only two categories: boards running Crashpad and boards
running Breakpad. Therefore, there is now only 1 USE flag:
force_breakpad. If force_breakpad is set, then the boards (scarlet, bob,
kevin, rainier, oak, elm, and hana) will continue to run only Breakpad;
otherwise, they will be at 100% Crashpad.

Remove the reference to force_crashpad and switch to using
force_breakpad to determine if Crashpad is running.

BUG=chromium:1038427
TEST=Ran on cyan, hana, kukui, and hana-kernelnext.
* Restarted Chrome multiple times. hana always used
  --no-enable-crashpad. All others used --enable-crashpad and had
  --crashpad-handler-pid
* Killed Chrome with SEGV. All board produced crash dumps
* Ran ChromeCrash* tast tests. hana skipped all Crashpad tests. All
  others skipped all Breakpad tests. All tests except
  ui.ChromeCrashLoggedIn.gpu_process_breakpad passed.

Cq-Depend: chromium:2335749, chromium:2334514, chromium:2335783, chromium:2335750, chrome-internal:3197841, chrome-internal:3198092
Change-Id: I2f451e0bf1060d1bc55acbd1b77c37896656c426