planner: Increase timeout for pre/post-test hooks.
Currently the timeout for pre/post-test hooks are 15s. It is no longer
long enough because we may delete large data files in test hooks.

In the longer term, we might want to allow test bundles to configure
test hook timeouts. But for now I'm just increasing the timeout to
mitigate crbug.com/1113573.

BUG=chromium:1113573
TEST=fast_build.sh -T

Change-Id: Ibf1dd514605a689f177577e20cdeacd19f62fd46