planner: Introduce fixtureStack.
fixtureStack is a key component to drive fixtures. It maintains
a stack of fixtures relevant to the current traversal state of a
fixture tree. The user of fixtureStack should just call Push, Pop,
Reset as it traverses a fixture tree; fixtureStack takes care of
calling fixture methods (SetUp, TearDown, Reset) appropriately.

BUG=chromium:1035940
TEST=fast_build.sh -T

Change-Id: I3bd0d42cf29c6efbd0008adc026b6683c463f38f