testing: Reorganize State types.
To prepare for introduction of fixtures, this patch reorganizes
State types. In the new world, there are three kinds of types
defined in state.go:

- Root types (EntityRoot, TestRoot): Keeps track of "states" shared
  among multiple State types. All State objects for an entity share
  the same Root object.
- State types (State, PreState, TestHookState, FixtState,
  FixtTestState): Provides API methods to user code that interact
  with the framework. Multiple State objects can be created for an
  entity (e.g. with State.Run).
- Mix-in types (globalMixin, testMixin): Actually implements API
  methods in State types. Since there are mutliple State types that
  provide similar but different sets of methods, they are implemented
  in mix-in types, and mix-in types are embedded to State types.

This patch is mostly a refactoring change. The only incompatible
change is that the Meta method is no longer exported in PreState and
TestHookState, but the method is not meant to be called by them from
the first place.

BUG=chromium:1035940
TEST=fast_build.sh -T

Change-Id: I92064139b8340a7360fcab4df20dca6c699264c0