testing: Panic on calling forbidden State methods.
Some State methods abort the test execution when they are called
inappropriately (e.g. DataPath for an undeclared data file). Formerly
Fatal/Fatalf were used in this case, but we are soon making a change
to refactor State types that makes it difficult to call Fatal/Fatalf
in some State methods.

This patch replaces those Fatal/Fatalf calls with panic.

This change should be not much visible to users since hitting these
code paths means programming errors in user code.

The only exception is a change to dataFS.Open. testing.dataFS had a
bug where it called s.DataPath for undeclared data files (it checked
whether it is safe to call s.DataPath by checking os.Stat for a file
path that would be returned by s.DataPath, but the check unexpectedly
passed if a file was not declared in the current test but it was in
another test). In this case we used to call s.Fatal on HTTP server
goroutines, which reports an error via testing.State and terminates
the goroutine with runtime.Goexit. I guess calling runtime.Goexit on
HTTP server goroutines is not really expected in general HTTP server
implementations, it somehow worked with httptest.Server. If we simply
replace the s.Fatal call in dataFS with panic, panic will abort the
whole process, not only the panicking goroutine. It is not desirable to
abort the test bundle only because unexpected file paths are requested
to the web server. For this reason, this patch fixes dataFS.Open to
correctly check whether a requested file is declared. This fix is done
atomically with s.Fatal -> panic change because it is difficult to do
the two changes separately.

BUG=chromium:1035940
TEST=fast_build.sh -T

Change-Id: I0ab5ea6c62520e5c513fbb5d1acc798b17ad2413