run: Separate non-exported fields from EntityResult.
EntityResult used to mix both exported fields (serialized to JSON) and
unexported fields used to keep track of states of a currently running
test, which was a bit tricky.

This change split it to two separate structs. entityState is a new type
that keeps track of states, and EntityResult is now a field of it.

Initially logFile is the only state field in entityState
(testStartMsgTime is removed since it is actually unused). Upcoming
changes will add a few more fields to the struct.

This is a refactoring change. No functional change is expected.

BUG=chromium:1035940
TEST=fast_build.sh -T

Change-Id: I5d53342d3a75624203466dc3c81bb3c59759fda2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2362246
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>