WIP: run: Support receiving multiple results for an entity.
We fixture support, we are making two protocol changes:

  1. Support receiving multiple results for an entity.
  2. Support multiple running entities at a time.

This is a change for (1).

With the introduction of fixtures, it is possible that Tast CLI
receives multiple TestStart/TestEnd messages for an identically named
entity (e.g. tests in two different test bundles use the same fixture).
Since currently Tast CLI assumes that an output directory for a test
$NAME is $OUTDIR/$NAME, if it receives multiple TestEnd messages for
identically named entities, it may overwrite existing directories.

To deal with this, this patch makes following changes:

- Test bundles create an intermediate test output directory with a
  random name to avoid name conflicts.
- Test bundles tell Tast CLI an intermediate output directory path by
  setting the new TestStart.OutDir field.
- Tast CLI copies output files from the directory indicated in
  TestStart.OutDir on receiving TestEnd. The destination directory is
  named $NAME for the first occurrence of an entity, and then $NAME.1,
  $NAME.2, and so on.

Note: This change modifies the protocol between Tast CLI and test
bundles, so we should trigger updating Tast CLI. However fixtures are
not generally available to normal users yet, and we will make further
changes to the protocol later anyway, so we are not triggering update
for now.

BUG=chromium:1035940
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=~/go/bin/tast run DUT example.Perf

Change-Id: Icb3b2f0c54635dfe695ee7d8def2693b433fe274