planner: Implement fixture execution.
This patch introduces fixtPlan to the planner which manages execution
plans of fixture-ready tests (i.e. tests not using preconditions).

planner.RunTests now takes a "start fixture" name in addition to
a list of tests. A start fixture specifies a node in a fixture tree
where the planner starts execution of fixtures. For conventional
local tests, the start fixture name is empty. Once we support remote
fixtures, a start fixture name will be that of a remote fixture set
up by remote_test_runner.

BUG=chromium:1035940
TEST=fast_build.sh -T

Change-Id: I983236177b4ab73095219385b658e24c54a07cd9