testing: add VarDeps support
VarDeps declares test dependencies on runtime variables.
VarDeps must be a subset of Vars, and it's checked in newTestInstance.
t.ShouldRun is updated to receive runtime variables, and decide to skip
tests when var dependency is not satisfied.
EntityInfo is also updated to transfer tests' VarDeps to runners.

Currently Params are not updated. We might want to have parameterized
variables, but for now they are not necessary.

BUG=chromium:1085320
TEST=fast_build.sh -T

Change-Id: Id240a39781068c473611ffeec3b1e78f9de203be