hwdeps: Create ChromeEC hwdep
The EmbeddedController topology was added to Boxster in
https://crrev.com/c/2382175, and ECType was added to Device Configs in
https://crrev.com/c/2375909.

This CL adds the EC feature to Tast, and creates a new HWDep to allow
tests to run iff their EC is a ChromeEC.

Note that the EC topology has not yet been populated for each board. As
a result, for now the HwDep will always yield no EC. See the bug for
backfilling the topology data at https://crbug.com/1139017.

BUG=b:163135224
TEST=fast_build.sh -T
TEST=Run a Tast test with HardwareDeps: hwdep.D(hwdep.ChromeEC())

Change-Id: I1a1fc985ac96370460e1626cecf683368ef939c3